Eliminate require rescue block; cd to script dir
Since this script lives in the go_script gem itself, there's little need to
check whether it exists or to install it.

Also borrowed an update from the template, whereby the script always executes
within its own directory, to ensure the expected behavior when invoking the
script from another directory.