Abort if go_script not in Gemfile
Found out that when adding a `./go` script to a project for the first time, if
that project has a Gemfile, the `./go` script will enter a loop whereby it
installs go_script to the system gems over and over. This change causes the
script to abort with instructions, since it's up to the project developer to
determine how exactly the go_script gem should be included (e.g. as part of a
'development' group).